CG→checks scale bridge: skills never reach checks, stats stored unscaled, Brewing unlinked (Apostate's ×10 ruling) (#2903)
CG-checks scale bridge (#2894): stats store x10, skills display true value, CG skills bridge to trait rows, Brewing linked

Apostate's ruling (ADR-0193): stats display single-digit but store x10 under
the hood; skills store AND display their true 1-100 value (single-point
development gains, XP unlocks at the x10 rung boundaries).

- _create_stat_values writes internal x10 (draft allocations stay 1-5 dots);
  _apply_post_cg_bonuses scales its dormant display-dot bonuses the same way
- _create_skill_values bridges each CG skill verbatim into a
  CharacterTraitValue row (perform_check reads only trait rows; DP level-ups
  now increment the CG row instead of get_or_create'ing a fresh one at 10)
- fatigue drops the dual-convention guard: display = value // STAT_DISPLAY_DIVISOR
- maturation spends +10 internal per display dot; caps stay authored in dots
  and convert at the comparison; the panel API ships display dots
- display seam is now type-aware via traits.models.display_trait_value:
  skill display paths stop dividing by 10 (model display_value properties,
  trait handler, thread-hub picker, unlock requirement messages, admin
  columns, telnet skill-breakthrough lines); stats keep the /10 edge
- provisioning seed links Brewing to the Cooking check via
  CheckTypeSpecialization (weight 1.0); the runtime-spec no-weight branch is
  documented as by-design (spec is fully additive)
- STAT_DISPLAY_DIVISOR canonical home: world.traits.models (re-exported via
  traits.constants and character_creation.constants)
- gen-api-types regen (weavable display_value double->integer); known local
  drift (consequence-pool id flip, action_category drop) hand-reverted
- ADR-0193; INDEX/character_creation docs + traits glossary updated
- acceptance tests: CG-finalized character == hand-seeded twin on check points

Closes #2894

+++ src/world/magic/serializers.py
@@ -2791,7 +2791,8 @@ class _WeavableTraitSerializer(serializers.Serializer):
     trait_id = serializers.IntegerField()
     name = serializers.CharField()
     trait_type = serializers.CharField()
-    display_value = serializers.FloatField()
+    # Stats ÷10, skills true value (#2894) — always whole numbers now.
+    display_value = serializers.IntegerField()
 
 
 class _WeavableTechniqueSerializer(serializers.Serializer):
